requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
email-validator==2.1.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0